+++ src/main/java/com/ecomerce/recursos/PedidoRecurso.java
@@ -27,7 +27,7 @@ public class PedidoRecurso {
 	}
 	
 	@GetMapping(value = "/{id}")
-	public ResponseEntity<Pedido> findById(@PathVariable Long id){
+	public ResponseEntity<Pedido> findById(@PathVariable Integer id){
 		Pedido obj = servico.findById(id);
 		return ResponseEntity.ok().body(obj);
 	}
